feat: Implementa endpoint de URL firmada para videos

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, TIMESTAMP, TEXT, INT, BIGINT, ForeignKey, JSON, Enum
+from sqlalchemy import Column, String, TIMESTAMP, TEXT, INT, INT, ForeignKey, JSON, Enum
 from sqlalchemy.dialects.mysql import TINYINT
 from sqlalchemy.sql import func
 from .database import Base  # heredamos la clase que definimos en database.py
@@ -33,7 +33,7 @@ class User(Base):
 
 class Module(Base):
     __tablename__ = "modules"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
+    id = Column(INT, primary_key=True, autoincrement=True)
     code = Column(String(64), unique=True)
     title = Column(String(120), nullable=False)
     description = Column(TEXT, nullable=True)
@@ -48,8 +48,8 @@ class Module(Base):
 
 class Lesson(Base):
     __tablename__ = "lessons"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
-    module_id = Column(BIGINT, ForeignKey("modules.id"), nullable=False)
+    id = Column(INT, primary_key=True, autoincrement=True)
+    module_id = Column(INT, ForeignKey("modules.id"), nullable=False)
     title = Column(String(120), nullable=False)
     sort_order = Column(INT, default=0)
     
@@ -60,7 +60,7 @@ class Lesson(Base):
 
 class Sign(Base):
     __tablename__ = "signs"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
+    id = Column(INT, primary_key=True, autoincrement=True)
     word = Column(String(120), nullable=False, index=True) # [cite: 192, 200]
     category = Column(String(80), nullable=True) # [cite: 193]
     video_path = Column(String(512), nullable=False) # [cite: 195]
@@ -74,8 +74,8 @@ class Sign(Base):
 
 class Quiz(Base):
     __tablename__ = "quizzes"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
-    module_id = Column(BIGINT, ForeignKey("modules.id"), nullable=False)
+    id = Column(INT, primary_key=True, autoincrement=True)
+    module_id = Column(INT, ForeignKey("modules.id"), nullable=False)
     type = Column(Enum('multiple_choice', 'complete', 'pair'), nullable=False) # [cite: 211]
     title = Column(String(160))
 
@@ -85,8 +85,8 @@ class Quiz(Base):
 
 class QuizQuestion(Base):
     __tablename__ = "quiz_questions"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
-    quiz_id = Column(BIGINT, ForeignKey("quizzes.id"), nullable=False)
+    id = Column(INT, primary_key=True, autoincrement=True)
+    quiz_id = Column(INT, ForeignKey("quizzes.id"), nullable=False)
     prompt = Column(TEXT, nullable=False) # [cite: 218]
     options = Column(JSON, nullable=True) # [cite: 220]
     answer = Column(String(255), nullable=True) # [cite: 222]
@@ -97,9 +97,9 @@ class QuizQuestion(Base):
 
 class SignPair(Base):
     __tablename__ = "sign_pairs"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
+    id = Column(INT, primary_key=True, autoincrement=True)
     word = Column(String(120), nullable=False) # [cite: 230]
-    sign_id = Column(BIGINT, ForeignKey("signs.id"), nullable=False) # [cite: 231]
+    sign_id = Column(INT, ForeignKey("signs.id"), nullable=False) # [cite: 231]
 
     sign = relationship("Sign", back_populates="sign_pairs")
 
@@ -108,7 +108,7 @@ class SignPair(Base):
 class UserModuleProgress(Base):
     __tablename__ = "user_module_progress"
     user_id = Column(String(128), ForeignKey("users.uid"), primary_key=True) # [cite: 238]
-    module_id = Column(BIGINT, ForeignKey("modules.id"), primary_key=True) # [cite: 240]
+    module_id = Column(INT, ForeignKey("modules.id"), primary_key=True) # [cite: 240]
     percent = Column(INT, nullable=False, default=0) # [cite: 242]
     last_activity = Column(TIMESTAMP(timezone=True), server_default=func.now(), onupdate=func.now()) # [cite: 244]
 
@@ -117,9 +117,9 @@ class UserModuleProgress(Base):
 
 class QuizAttempt(Base):
     __tablename__ = "quiz_attempts"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
+    id = Column(INT, primary_key=True, autoincrement=True)
     user_id = Column(String(128), ForeignKey("users.uid"), nullable=False, index=True) # [cite: 253]
-    quiz_id = Column(BIGINT, ForeignKey("quizzes.id"), nullable=False) # [cite: 255]
+    quiz_id = Column(INT, ForeignKey("quizzes.id"), nullable=False) # [cite: 255]
     score = Column(INT, nullable=False) # [cite: 257]
     total = Column(INT, nullable=False) # [cite: 259]
     duration_ms = Column(INT, nullable=True) # [cite: 261]
@@ -130,9 +130,9 @@ class QuizAttempt(Base):
 
 class MemoryRun(Base):
     __tablename__ = "memory_runs"
-    id = Column(BIGINT, primary_key=True, autoincrement=True)
+    id = Column(INT, primary_key=True, autoincrement=True)
     user_id = Column(String(128), ForeignKey("users.uid"), nullable=False) # [cite: 269]
-    module_id = Column(BIGINT, ForeignKey("modules.id"), nullable=True) # [cite: 271]
+    module_id = Column(INT, ForeignKey("modules.id"), nullable=True) # [cite: 271]
     matches = Column(INT) # [cite: 273]
     attempts = Column(INT) # [cite: 275]
     streak = Column(INT) # [cite: 277]
